+++ projects/journey-maps-client/src/lib/components/info-box/info-box.component.ts
@@ -13,11 +13,11 @@ import {LocaleService} from '../../services/locale.service';
   animations: [
     trigger('showHideTrigger', [
       transition(':enter', [
-        style({transform: 'translateX(100%)'}),
+        style({transform: 'translateX(-100%)'}),
         animate('500ms', style({transform: 'translateX(0%)'})),
       ]),
       transition(':leave', [
-        animate('500ms', style({transform: 'translateX(100%)'}))
+        animate('500ms', style({transform: 'translateX(-100%)'}))
       ])
     ]),
   ]
@@ -45,7 +45,10 @@ export class InfoBoxComponent implements OnInit {
   }
 
   shouldRender(): boolean {
-    return !!this.selectedMarker;
+    return !!this.selectedMarker && (
+      !!this.selectedMarker.infoBlocks?.length ||
+      !!this.infoBoxTemplate
+    );
   }
 
   getInfoBlocks(): InfoBlock[] {