+++ projects/journey-maps-client/src/lib/model/marker.ts
@@ -33,6 +33,18 @@ export interface Marker {
   subtitle?: string;
   /**
    * List of info blocks that will be displayed in the overlay when the marker is selected.
+   *
+   * Unless you have defined a {@link JourneyMapsClientComponent#infoBoxTemplate},
+   * the overlay will only appear if this field is set and contains at least one infoBlock.
    */
   infoBlocks?: InfoBlock[];
+  /**
+   * Url that will be opened if the marker is selected instead of the overlay.
+   */
+  markerUrl?: string;
+  /**
+   * If true or undefined the module will emit an event in output parameter selectedMarkerId:string when this marker is selected.
+   * Default value is true.
+   */
+  triggerEvent?: boolean;
 }